+++ api/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2024-07-13
+
+### Added
+
+- Add coerce in the POST hydration body schema
+- Add id in the GET hydration response object
+
+## [1.0.1] - 2024-07-13
+
+### Fixed
+
+- Add CORS policy that to allow all origins access API
+
 ## [1.0.0] - 2024-07-13
 
 ### Added